requirements.txt
wheel
setuptools
numpy>=1.23.5
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-openai>=0.0.2
pydantic>=2.5.0
python-dotenv
faiss-cpu
openai>=1.6.1
httpx>=0.24.1
gradio>=4.19.1
python-docx>=0.8.11
unstructured>=0.10.30
unstructured[docx]>=0.10.30
python-magic-bin>=0.4.14; platform_system == 'Windows'
python-magic>=0.4.27; platform_system != 'Windows' 